Implemented new modulerisation of schema.
